Update LineBreaker_Delegate following commit d5621ed
Change Ifa95f870 changed the native API of LineBreaker.java. This
updates the corresponding delegate.

Bug: 74062470
Test: layoutlib tests

+++ bridge/src/android/graphics/text/LineBreaker_Delegate.java
@@ -91,8 +91,8 @@ public class LineBreaker_Delegate {
             float firstWidth,
             int firstWidthLineCount,
             float restWidth,
-            @Nullable int[] variableTabStops,
-            int defaultTabStop,
+            @Nullable float[] variableTabStops,
+            float defaultTabStop,
             int indentsOffset) {
         Builder builder = sBuilderManager.getDelegate(nativePtr);
         if (builder == null) {
